Sistema de buscar/alteração/remoção concluído

+++ menus/pesquisa_exibir_aluno.py
@@ -13,7 +13,6 @@ import platform
 import getpass
 from datetime import *
 
-
 arqAlunos = './data/database.genius'
 
 def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoFechar):
@@ -27,6 +26,9 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
 
     s = ttk.Style()
     s.configure('branco.TFrame',background='white')
+
+    t = ttk.Style()
+    t.configure('branco.TLabelframe',font=('Segoe UI',12,'normal'),background='white',foreground='white')
 
     version = 'Genius (LITE) - 7.0.1'
     altura       = 5
@@ -51,6 +53,7 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     
     #Box Alunos
     labelBox = ttk.Labelframe(frameGeral,text='ALUNOS ENCONTRADOS NO BANCO DE DADOS',width=largura2)
+    labelBox['style'] = 'branco.TLabelframe'
     labelBox.grid(row=1,column=0,sticky=W+E)
     
     global boxAlunos
@@ -66,32 +69,47 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     #Suporte1
     suporte1 = ttk.Label(frameGeral,text='')
     suporte1.grid(row=2,column=0,sticky=W+E)
+    suporte1['style'] = 'branco.TLabel'
 
     #Frame para botões de filtro
     filtros = ttk.Frame(frameGeral)
     filtros.grid(row=3,column=0,sticky=W+E)
+    filtros['style'] = 'branco.TFrame'
 
-    btAlfab = ttk.Button(filtros,text='Filtrar por Nome')
-    btAlfab.grid(row=0,column=0)
+    imageAlfab = ImageTk.PhotoImage(Image.open('./icons/30/AlphabeticalSorting30px.png'))
+    btAlfab = ttk.Button(filtros,text='Filtrar por ordem alfabética',compound=TOP,image=imageAlfab,command=partial(organizaAlunosNome,boxAlunos,dictAlunos))
+    btAlfab.grid(row=0,column=0,ipady=altura)
+    btAlfab.image = imageAlfab
 
-    btTurma = ttk.Button(filtros,text='Filtrar por Turma')
-    btTurma.grid(row=0,column=1)
+    imageTurma = ImageTk.PhotoImage(Image.open('./icons/30/Numeric30px.png'))
+    btTurma = ttk.Button(filtros,text='Filtrar por Turma',compound=TOP,image=imageTurma,command=partial(organizaAlunosTurma,boxAlunos,dictAlunos))
+    btTurma.grid(row=0,column=1,ipady=altura)
+    btTurma.image = imageTurma
+
+    imageChave = ImageTk.PhotoImage(Image.open('./icons/30/Access30px.png'))
+    btChave = ttk.Button(filtros,text='Filtrar por chave',compound=TOP,image=imageChave,command=partial(organizaAlunosChave,boxAlunos,dictAlunos))
+    btChave.grid(row=0,column=2,ipady=altura)
+    btChave.image = imageChave
 
     #Suporte2
     suporte2 = ttk.Label(frameGeral,text='')
     suporte2.grid(row=4,column=0,sticky=W+E)
+    suporte2['style'] = 'branco.TLabel'
 
     #Label para edição de campos
     editLabel = ttk.Labelframe(frameGeral,text='MODIFICAR INFORMAÇÕES',width=largura2)
+    editLabel['style'] = 'branco.TLabelframe'
     editLabel.grid(row=6,column=0,sticky=W)
 
     #Frame contendo os campos para edição
     frameEdit = ttk.Frame(editLabel)
     frameEdit.grid(row=0,column=0,columnspan=9,sticky=W+E)
+    frameEdit['style'] = 'branco.TLabel'
 
     #Nome
     nome = ttk.Label(frameEdit,text='NOME',font=fonteTopo3)
     nome.grid(row=0,column=0,sticky=W)
+    nome['style'] = 'branco.TLabel'
 
     global inNome
     inNome = ttk.Entry(frameEdit, font=fonteTopo2)
@@ -100,14 +118,20 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     #Código
     codigo = ttk.Label(frameEdit,text='CÓDIGO',font=fonteTopo3)
     codigo.grid(row=2,column=0,sticky=W)
+    codigo['style'] = 'branco.TLabel'
 
     global inCodigo
     inCodigo = ttk.Entry(frameEdit, font=fonteTopo2)
     inCodigo.grid(row=3,column=0,sticky=W+E)
 
+    #MEIO1
+    meio1 = ttk.Label(frameEdit,text='')
+    meio1.grid(row=3,column=4)
+    meio1['style'] = 'branco.TLabel'
     #Turno
     turno = ttk.Label(frameEdit,text='TURNO',font=fonteTopo3)
     turno.grid(row=2,column=5,sticky=W)
+    turno['style'] = 'branco.TLabel'
 
     global inTurno
     inTurno = ttk.Combobox(frameEdit, font=fonteTopo2)
@@ -116,9 +140,15 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     inTurno.current(0)
     inTurno.grid(row=3,column=5,sticky=W+E)
 
+    #MEIO2
+    meio2 = ttk.Label(frameEdit,text='')
+    meio2.grid(row=3,column=9)
+    meio2['style'] = 'branco.TLabel'
+
     #Turma
     turma = ttk.Label(frameEdit,text='TURMA',font=fonteTopo3)
     turma.grid(row=2,column=10,sticky=W)
+    turma['style'] = 'branco.TLabel'
 
     global inTurma
     inTurma = ttk.Combobox(frameEdit, font=fonteTopo2)
@@ -130,6 +160,7 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     #Ano
     ano = ttk.Label(frameEdit,text='ANO',font=fonteTopo3)
     ano.grid(row=6,column=0,sticky=W)
+    ano['style'] = 'branco.TLabel'
 
     global inAno
     inAno = ttk.Spinbox(frameEdit, font=fonteTopo2)
@@ -138,6 +169,7 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     #Status A/N
     status = ttk.Label(frameEdit,text='STATUS',font=fonteTopo3)
     status.grid(row=8,column=0,sticky=W)
+    status['style'] = 'branco.TLabel'
 
     global inStatus
     inStatus = ttk.Combobox(frameEdit, font=fonteTopo2)
@@ -146,26 +178,42 @@ def pesquisaAluno(login,nivel,usuario,dictAlunos,janelaPai,janelaFechar, funcaoF
     #Suporte3
     suporte3 = ttk.Label(frameGeral,text='')
     suporte3.grid(row=7,column=0,sticky=W+E)
+    suporte3['style'] = 'branco.TLabel'
 
     #Frame para agrupar os botões 
     botoes = ttk.Frame(frameGeral)
     botoes.grid(row=8,column=0,sticky=W+S+N+E)
+    botoes['style'] = 'branco.TFrame'
 
-    btImprimir = ttk.Button(botoes,text='Imprimir')
-    btImprimir.grid(row=0,column=0)
+    imagePrint = ImageTk.PhotoImage(Image.open('./icons/30/Print30px.png'))
+    btImprimir = ttk.Button(botoes,text='Imprimir',width=largura,compound=TOP,image=imagePrint,command=partial(imprimeUnico,boxAlunos,dictAlunos,inCodigo))
+    btImprimir.grid(row=0,column=0,ipady=altura)
+    btImprimir.image = imagePrint
     
-    btSalvar = ttk.Button(botoes,text='Salvar',command=partial(ponte,inCodigo,inNome,inTurma,inTurno,inAno,login,dictAlunos,1))
-    btSalvar.grid(row=0,column=1)
+    imagePrintTudo = ImageTk.PhotoImage(Image.open('./icons/30/SendtoPrinter30px.png'))
+    btImprimirTudo = ttk.Button(botoes,text='Imprimir Tudo',width=largura,compound=TOP,image=imagePrintTudo,command=partial(imprimeGeral,boxAlunos,dictAlunos))
+    btImprimirTudo.grid(row=0,column=1,ipady=altura)
+    btImprimirTudo.image = imagePrintTudo
+    
+    imageSalvar = ImageTk.PhotoImage(Image.open('./icons/30/Save30px.png'))
+    btSalvar = ttk.Button(botoes,text='Salvar',width=largura,compound=TOP,image=imageSalvar,command=partial(ponte,inCodigo,inNome,inTurma,inTurno,inAno,login,dictAlunos,1))
+    btSalvar.grid(row=0,column=2,ipady=altura)
+    btSalvar.image = imageSalvar
 
-    btExcluir = ttk.Button(botoes,text='Excluir',command=partial(ponte,inCodigo,inNome,inTurma,inTurno,inAno,login,dictAlunos,2))
-    btExcluir.grid(row=0,column=2)
+    imageExcluir = ImageTk.PhotoImage(Image.open('./icons/30/TrashCan30px.png'))
+    btExcluir = ttk.Button(botoes,text='Excluir',width=largura,compound=TOP,image=imageExcluir,command=partial(ponte,inCodigo,inNome,inTurma,inTurno,inAno,login,dictAlunos,2))
+    btExcluir.grid(row=0,column=3,ipady=altura)
+    btExcluir.image = imageExcluir
 
-    btVoltar = ttk.Button(botoes,text='Voltar',command=partial(funcaoFechar,login,usuario,nivel,dictAlunos,janelaPai,frameGeral))
-    btVoltar.grid(row=0,column=3)
+    imageVoltar = ImageTk.PhotoImage(Image.open('./icons/30/Cancel30px.png'))
+    btVoltar = ttk.Button(botoes,text='Voltar',width=largura,compound=TOP,image=imageVoltar,command=partial(funcaoFechar,login,usuario,nivel,dictAlunos,janelaPai,frameGeral))
+    btVoltar.grid(row=0,column=4,ipady=altura)
+    btVoltar.image = imageVoltar
 
     #Suporte4
     suporte4 = ttk.Label(frameGeral,text='')
     suporte4.grid(row=9,column=0,sticky=W+E)
+    suporte4['style'] = 'branco.TLabel'
     
 def carregaParaEdicao(evt):
     '''
@@ -178,8 +226,12 @@ def carregaParaEdicao(evt):
 
     alunos = {}
     carregaAlunosArquivo(arqAlunos,alunos)
-    codAluno = valor2[0]
-    infoCompleta = alunos[codAluno]
+    try:
+        codAluno = valor2[0]
+        infoCompleta = alunos[codAluno]
+    except:
+        codAluno = valor2[-1]
+        infoCompleta = alunos[codAluno]
 
     colocaFrame(infoCompleta)
 
@@ -237,10 +289,13 @@ def ponte(inCodigo,inNome,inTurma,inTurno,inAno,cpf,dictAlunos,opcao):
     ano = inAno.get()
     ano = ano.upper()
 
-    if opcao == 1:
-        salvarModificacoesAlunos(codigo,nome,turma,turno,ano,cpf,dictAlunos)
-    elif opcao == 2:
-        excluirAluno(codigo,nome,turma,turno,ano,cpf,dictAlunos)
-        limpa()
-    boxAlunos.delete(0,END)
-    carregaAlunosBox(boxAlunos,dictAlunos)
+    if codigo == '':
+        messagebox.showerror('Ops!','Aluno não selecionado!')
+    else:
+        if opcao == 1:
+            salvarModificacoesAlunos(codigo,nome,turma,turno,ano,cpf,dictAlunos)
+        elif opcao == 2:
+            excluirAluno(codigo,nome,turma,turno,ano,cpf,dictAlunos)
+            limpa()
+        boxAlunos.delete(0,END)
+        carregaAlunosBox(boxAlunos,dictAlunos)
